+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - changed : center vertically datatables rows content
 - fixed : tabs body overflow hidden when not datatables
 - fixed : extendable text in tab info default 2 lines visible
+- removed : default bulma css
 
 ## 0.9.0 (2024-10-23)